modularized header, finalizing design
