+++ WriterTyper/client/src/components/SHARED/Footer/Footer.css
@@ -3,5 +3,5 @@ height: 10vh;
 bottom: 0px;
 left: 0;
 right: 0;
-background-color: #E0E0E0;
+background-color: #74888a;
 }
