+++ WriterTyper/client/src/components/SHARED/Header/Header.css
@@ -1,33 +1,11 @@
-@import url("https://use.typekit.net/yns4wlf.css");
 header {
   margin-top: -16px;
   height: 11.5vh;
   top: 0px;
   left: 0;
   right: 0;
-  background-color: #E0E0E0;
+  background-color: #74888a;
   display: flex;
   flex-flow: row;
   align-items: center;
-}
-.logo {
-  font-family: itc-american-typewriter, serif;
-  font-weight: 700;
-  font-style: italic;
-  font-size: 50px;
-  margin-left: 100px;
-}
-.navCont{
-  display: flex;
-  margin-left: 400px;
-  align-items: center;
-}
-.navItem {
-  display: inline;
-  text-decoration: none;
-  padding-right: 55px;
-  font-size: 40px;
-  font-family: itc-american-typewriter-cond, serif;
-  font-weight: 500;
-  font-style: normal;
 }
