+++ WriterTyper/client/src/components/SHARED/Header/Logo/Logo.css
@@ -1,0 +1,8 @@
+@import url("https://use.typekit.net/yns4wlf.css");
+.logo {
+  font-family: itc-american-typewriter, serif;
+  font-weight: 700;
+  font-style: italic;
+  font-size: 50px;
+  margin-left: 100px;
+}
