+++ WriterTyper/client/src/components/SHARED/Header/Nav/Nav.css
@@ -1,0 +1,15 @@
+@import url("https://use.typekit.net/yns4wlf.css");
+.navCont{
+  display: flex;
+  margin-left: 400px;
+  align-items: center;
+}
+.navItem {
+  display: inline;
+  text-decoration: none;
+  padding-right: 55px;
+  font-size: 40px;
+  font-family: itc-american-typewriter-cond, serif;
+  font-weight: 500;
+  font-style: normal;
+}
