+++ WriterTyper/client/src/components/SHARED/Header/Nav/NavButtonDesign/Button.css
@@ -1,0 +1,5 @@
+@import url("https://use.typekit.net/yns4wlf.css");
+.navButton {
+  background: transparent;
+  border: none;
+}
