+++ WriterTyper/client/src/components/SHARED/Header/Nav/Read/Read.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import { Link } from 'react-router-dom'
+import '../NavButtonDesign/Button.css'
+
+function ReadButton() {
+  return (
+    <Link to='/read' className='navLink'>
+      <button className='navButton'>
+        Read
+      </button>
+    </Link>
+  );
+}
+export default ReadButton;
