+++ WriterTyper/client/src/components/SHARED/Header/Nav/Upload/Upload.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import { Link } from 'react-router-dom'
+import '../NavButtonDesign/Button.css'
+
+function UploadButton() {
+  return (
+    <Link to='/upload' className='navLink'>
+      <button className='navButton'>
+        Upload
+      </button>
+    </Link>
+  );
+}
+export default UploadButton;
