--- WriterTyper/client/src/components/SHARED/Header/Read/Read.css
@@ -1,3 +1,0 @@
-.navLink {
-  text-decoration: none;
-}
