--- WriterTyper/client/src/components/SHARED/Header/Upload/Upload.css
@@ -1,3 +1,0 @@
-.navLink {
-  text-decoration: none;
-}
